requirements.txt
openai==1.3.6
pinecone-client==3.0.2
python-dotenv==1.0.0
Flask==2.3.3
python-docx==0.8.11
flask-cors==3.0.10
flask-limiter==3.5.0